+++ plugins/mariadb/versions/11.0/install.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-PATH=/bin:/sbin:/usr/bin:/usr/sbin:/usr/local/bin:/usr/local/sbin:~/bin
+PATH=/bin:/sbin:/usr/bin:/usr/sbin:/usr/local/bin:/usr/local/sbin:~/bin:/opt/homebrew/bin
 export PATH
 
 #https://dev.mysql.com/downloads/mysql/5.5.html#downloads
@@ -14,7 +14,7 @@ sysName=`uname`
 install_tmp=${rootPath}/tmp/mw_install.pl
 mariadbDir=${serverPath}/source/mariadb
 
-MY_VER=11.0.1
+MY_VER=11.0.3
 
 Install_app()
 {
@@ -101,16 +101,18 @@ Install_app()
 
 		if [ -d $serverPath/mariadb ];then
 			echo '10.11' > $serverPath/mariadb/version.pl
-			echo '安装完成' > $install_tmp
+			echo '安装完成'
 		else
-			echo '安装失败' > $install_tmp
+			echo '安装失败'
 			echo 'install fail'>&2
 			exit 1
 		fi
 	fi
 
-	rm -rf ${mariadbDir}/mariadb-${MY_VER}
-	rm -rf ${mariadbDir}/mariadb-${MY_VER}.tar.gz
+	if [ -d ${mariadbDir}/mariadb-${MY_VER} ];then
+		rm -rf ${mariadbDir}/mariadb-${MY_VER}
+	fi
+
 }
 
 Uninstall_app()
